Renamed combined.rs to root_and_tab_data.rs
