+++ nushift/src/model/mod.rs
@@ -3,10 +3,10 @@
 
 pub(crate) mod root_data;
 pub(crate) mod tab_data;
-mod combined;
+mod root_and_tab_data;
 pub(crate) mod scale_and_size;
 pub(crate) mod client_framebuffer;
 
 pub use root_data::RootData;
 pub use tab_data::TabData;
-pub use combined::RootAndTabData;
+pub use root_and_tab_data::RootAndTabData;
